+++ cc/hw/disk/device_kbd.64.c
@@ -3,7 +3,7 @@
 #include <cpu.h>
 #include <apic.h>
 
-void dev_kbd_isr(interrupt_frame_t* frame, uint16_t intnum){
+void dev_kbd_isr(interrupt_frame_t* frame, uint8_t intnum){
 	UNUSED(frame);
 	printf("KEYBOARD: Info keyboard event occured at %i\n", intnum);
